requirements.txt
tweepy
selenium
Pillow
psutil
requests
mastodon.py
discord-webhook
mysqlclient
